requirements.txt
coverage==3.7.1
cssmin==0.2.0
Flask==0.10.1
Flask-Assets==0.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
nose==1.3.7
nosexcover==1.0.10
PyYAML==3.11
webassets==0.11.1
Werkzeug==0.10.4
wheel==0.24.0
